// PI_FIRE_BACKEND-master/src/types/ocorrencias.ts
export interface Ocorrencia {
  id?: number;
  foto : string | null;
  carimbo_data_hora?: Date;
  numero_aviso: string;
  diretoria: string;
  grupamento: string;
  ponto_base: string;
  data_acionamento: Date;
  natureza_ocorrencia: string;
  grupo_ocorrencia: string;
  subgrupo_ocorrencia: string;
  situacao_ocorrencia: string;
  ocorrencia_nao_atendida?: boolean;
  horario_saida_quartel?: string;
  horario_chegada_local?: string;
  motivo_nao_atendida?: string;
  motivo_sem_atuacao?: string;
  horario_saida_local?: string;
  vitima_envolvida?: boolean;
  sexo_vitima?: 'Masculino' | 'Feminino' | 'Outro';
  idade_vitima?: number;
  classificacao_vitima?: string;
  destino_vitima?: string;
  viatura_empregada: string;
  numero_viatura: string;
  forma_acionamento: string;
  local_acionamento: string;
  municipio: string;
  regiao: string;
  bairro: string;
  tipo_logradouro: string;
  ais: string;
  logradouro: string;
  latitude?: number;
  longitude?: number;
  created_at?: Date;
  updated_at?: Date;
}

export type CreateOcorrencia = Omit<Ocorrencia, 'id' | 'created_at' | 'updated_at'>;